Fix up types.

+++ TJImageCache/TJImageCache.h
@@ -41,7 +41,7 @@ NS_ASSUME_NONNULL_BEGIN
 + (void)removeImageAtURL:(NSString *const)url;
 + (void)dumpDiskCache;
 + (void)dumpMemoryCache;
-+ (void)getDiskCacheSize:(void (^const)(NSUInteger diskCacheSize))completion;
++ (void)getDiskCacheSize:(void (^const)(long long diskCacheSize))completion;
 
 + (void)auditCacheWithBlock:(BOOL (^const)(NSString *hashedURL, NSDate *lastAccess, NSDate *createdDate))block completionBlock:(void (^_Nullable)(void))completionBlock; // return YES to preserve the image, return NO to delete it
 + (void)auditCacheWithBlock:(BOOL (^const)(NSString *hashedURL, NSDate *lastAccess, NSDate *createdDate))block;
